Refactor event-related handlers and routers
